feat: 음악 공유 기능 추가

+++ src/main/java/io/github/eappezo/soundary/services/friend/infrastructure/QFriendChecker.java
@@ -37,7 +37,7 @@ public class QFriendChecker implements FriendChecker {
                 fromTable.toUserId.in(rawFriendIds)
             )
             .fetch()
-            .getFirst().intValue() == rawFriendIds.size();
+            .getFirst() == rawFriendIds.size();
     }
